chore(command): minor cleanup
